Move torch.utils._benchmark to torch.utils.benchmark. Yay!
