+++ benchmarks/record_function_benchmark/record_function_bench.py
@@ -1,7 +1,7 @@
 import argparse
 import sys
 import torch
-import torch.utils._benchmark as benchmark_utils
+import torch.utils.benchmark as benchmark_utils
 
 
 try:
