--- torch/utils/_benchmark/__init__.py
@@ -1,4 +1,0 @@
-from torch.utils._benchmark.utils.common import *
-from torch.utils._benchmark.utils.timer import *
-from torch.utils._benchmark.utils.compare import *
-from torch.utils._benchmark.utils.fuzzer import *
