+++ torch/utils/benchmark/__init__.py
@@ -1,0 +1,4 @@
+from torch.utils.benchmark.utils.common import *
+from torch.utils.benchmark.utils.timer import *
+from torch.utils.benchmark.utils.compare import *
+from torch.utils.benchmark.utils.fuzzer import *
