+++ torch/utils/benchmark/op_fuzzers/binary.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch
 
-from torch.utils._benchmark import Fuzzer, FuzzedParameter, ParameterAlias, FuzzedTensor
+from torch.utils.benchmark import Fuzzer, FuzzedParameter, ParameterAlias, FuzzedTensor
 
 
 _MIN_DIM_SIZE = 16
@@ -12,21 +12,23 @@ _POW_TWO_SIZES = tuple(2 ** i for i in range(
 ))
 
 
-class UnaryOpFuzzer(Fuzzer):
+class BinaryOpFuzzer(Fuzzer):
     def __init__(self, seed, dtype=torch.float32, cuda=False):
         super().__init__(
             parameters=[
-                # Dimensionality of x. (e.g. 1D, 2D, or 3D.)
+                # Dimensionality of x and y. (e.g. 1D, 2D, or 3D.)
                 FuzzedParameter("dim", distribution={1: 0.3, 2: 0.4, 3: 0.3}, strict=True),
 
-                # Shapes for `x`.
-                #   It is important to test all shapes, however
+                # Shapes for `x` and `y`.
+                #       It is important to test all shapes, however
                 #   powers of two are especially important and therefore
                 #   warrant special attention. This is done by generating
                 #   both a value drawn from all integers between the min and
                 #   max allowed values, and another from only the powers of two
                 #   (both distributions are loguniform) and then randomly
                 #   selecting between the two.
+                #       Moreover, `y` will occasionally have singleton
+                #   dimensions in order to test broadcasting.
                 [
                     FuzzedParameter(
                         name=f"k_any_{i}",
@@ -52,12 +54,25 @@ class UnaryOpFuzzer(Fuzzer):
                     ) for i in range(3)
                 ],
 
-                # Steps for `x`. (Benchmarks strided memory access.)
                 [
                     FuzzedParameter(
-                        name=f"x_step_{i}",
-                        distribution={1: 0.8, 2: 0.06, 4: 0.06, 8: 0.04, 16: 0.04},
+                        name=f"y_k{i}",
+                        distribution={
+                            ParameterAlias(f"k{i}"): 0.8,
+                            1: 0.2,
+                        },
+                        strict=True,
                     ) for i in range(3)
+                ],
+
+                # Steps for `x` and `y`. (Benchmarks strided memory access.)
+                [
+                    FuzzedParameter(
+                        name=f"{name}_step_{i}",
+                        distribution={1: 0.8, 2: 0.06, 4: 0.06, 8: 0.04, 16: 0.04},
+                    )
+                    for i in range(3)
+                    for name in ("x", "y")
                 ],
 
                 # Repeatable entropy for downstream applications.
@@ -71,6 +86,16 @@ class UnaryOpFuzzer(Fuzzer):
                     probability_contiguous=0.75,
                     min_elements=4 * 1024,
                     max_elements=32 * 1024 ** 2,
+                    max_allocation_bytes=2 * 1024**3,  # 2 GB
+                    dim_parameter="dim",
+                    dtype=dtype,
+                    cuda=cuda,
+                ),
+                FuzzedTensor(
+                    name="y",
+                    size=("y_k0", "y_k1", "y_k2"),
+                    steps=("x_step_0", "x_step_1", "x_step_2"),
+                    probability_contiguous=0.75,
                     max_allocation_bytes=2 * 1024**3,  # 2 GB
                     dim_parameter="dim",
                     dtype=dtype,
