+++ torch/utils/benchmark/utils/common.py
